refactor: Reuse person to card presenter
This refactor re-uses the person to card component presenter in the
move to card component presenter as they shared a lot of logic and
we also want to ensure that if we're changing how the person is
presented in the card component, it is consistent across the dashboard
and anywhere else we use it.
