+++ common/presenters/person-to-card-component.js
@@ -2,38 +2,54 @@ const { filter } = require('lodash')
 
 const i18n = require('../../config/i18n')
 const filters = require('../../config/nunjucks/filters')
+const assessmentToTagList = require('./assessment-to-tag-list')
 
-function personToCardComponent({
-  gender = {},
-  fullname = '',
-  image_url: imageUrl,
-  date_of_birth: dateOfBirth,
-}) {
-  const dateOfBirthLabel = i18n.t('age', {
-    context: 'with_date_of_birth',
-    age: filters.calculateAge(dateOfBirth),
-    date_of_birth: filters.formatDate(dateOfBirth),
-  })
-  const metaItems = [
-    {
-      label: i18n.t('fields::date_of_birth.label'),
-      text: dateOfBirth ? dateOfBirthLabel : '',
-    },
-    {
-      label: i18n.t('fields::gender.label'),
-      text: gender.title,
-    },
-  ]
+function personToCardComponent({ showMeta = true, showTags = true } = {}) {
+  return function item({
+    href,
+    gender = {},
+    fullname = '',
+    image_url: imageUrl,
+    date_of_birth: dateOfBirth,
+    assessment_answers: assessmentAnswers,
+  }) {
+    const meta = {}
+    const tags = {}
 
-  return {
-    image_path: imageUrl,
-    image_alt: fullname.toUpperCase(),
-    title: {
-      text: fullname.toUpperCase(),
-    },
-    meta: {
-      items: filter(metaItems, 'text'),
-    },
+    if (showMeta) {
+      const dateOfBirthLabel = i18n.t('age', {
+        context: 'with_date_of_birth',
+        age: filters.calculateAge(dateOfBirth),
+        date_of_birth: filters.formatDate(dateOfBirth),
+      })
+      const metaItems = [
+        {
+          label: i18n.t('fields::date_of_birth.label'),
+          text: dateOfBirth ? dateOfBirthLabel : '',
+        },
+        {
+          label: i18n.t('fields::gender.label'),
+          text: gender.title,
+        },
+      ]
+
+      meta.items = filter(metaItems, 'text')
+    }
+
+    if (showTags) {
+      tags.items = assessmentToTagList(assessmentAnswers, href)
+    }
+
+    return {
+      href,
+      meta,
+      tags,
+      image_path: imageUrl,
+      image_alt: fullname.toUpperCase(),
+      title: {
+        text: fullname.toUpperCase(),
+      },
+    }
   }
 }
