+++ common/presenters/person-to-card-component.test.js
@@ -4,12 +4,35 @@ const i18n = require('../../config/i18n')
 const filters = require('../../config/nunjucks/filters')
 
 const mockPerson = {
+  href: '/move/12345',
   fullname: 'Name, Full',
   image_url: '/path/to/image.jpg',
   date_of_birth: '2000-10-10',
   gender: {
     title: 'Male',
   },
+  assessment_answers: [
+    {
+      key: 'concealed_items',
+      title: 'Concealed items',
+      category: 'risk',
+    },
+    {
+      key: 'health_issue',
+      title: 'Health issue',
+      category: 'health',
+    },
+    {
+      key: 'other_risks',
+      title: 'Any other risks',
+      category: 'risk',
+    },
+    {
+      key: 'legal_representation',
+      title: 'Solicitor or other legal representation',
+      category: 'court',
+    },
+  ],
 }
 
 describe('Presenters', function() {
@@ -25,10 +48,15 @@ describe('Presenters', function() {
     context('with default options', function() {
       context('with mock person', function() {
         beforeEach(function() {
-          transformedResponse = personToCardComponent(mockPerson)
+          transformedResponse = personToCardComponent()(mockPerson)
         })
 
         describe('response', function() {
+          it('should contain a href', function() {
+            expect(transformedResponse).to.have.property('href')
+            expect(transformedResponse.href).to.equal(mockPerson.href)
+          })
+
           it('should contain a title', function() {
             expect(transformedResponse).to.have.property('title')
             expect(transformedResponse.title).to.deep.equal({
@@ -61,6 +89,32 @@ describe('Presenters', function() {
                 {
                   label: '__translated__',
                   text: mockPerson.gender.title,
+                },
+              ],
+            })
+          })
+
+          it('should contain correct tags', function() {
+            expect(transformedResponse).to.have.property('tags')
+            expect(transformedResponse.tags).to.deep.equal({
+              items: [
+                {
+                  href: '/move/12345#concealed-items',
+                  text: 'Concealed items',
+                  classes: 'app-tag--destructive',
+                  sortOrder: 1,
+                },
+                {
+                  href: '/move/12345#any-other-risks',
+                  text: 'Any other risks',
+                  classes: 'app-tag--destructive',
+                  sortOrder: 1,
+                },
+                {
+                  href: '/move/12345#health-issue',
+                  text: 'Health issue',
+                  classes: '',
+                  sortOrder: 2,
                 },
               ],
             })
@@ -98,7 +152,7 @@ describe('Presenters', function() {
         let transformedResponse
 
         beforeEach(function() {
-          transformedResponse = personToCardComponent({
+          transformedResponse = personToCardComponent()({
             last_name: 'Jones',
             first_names: 'Steve',
             date_of_birth: '',
@@ -118,7 +172,7 @@ describe('Presenters', function() {
         let transformedResponse
 
         beforeEach(function() {
-          transformedResponse = personToCardComponent({
+          transformedResponse = personToCardComponent()({
             last_name: 'Jones',
             first_names: 'Steve',
             date_of_birth: '',
@@ -140,6 +194,178 @@ describe('Presenters', function() {
             ],
           })
         })
+      })
+
+      context('with assessment answers', function() {
+        let transformedResponse, mockPersonWithAnswers
+
+        beforeEach(function() {
+          mockPersonWithAnswers = {
+            href: '/move/12345',
+            last_name: 'Jones',
+            first_names: 'Steve',
+            date_of_birth: '',
+            assessment_answers: [
+              {
+                key: 'concealed_items',
+                title: 'Concealed items',
+                comments: 'Penknife found in trouser pockets',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: '942a634a-2e38-49be-bfe6-ac03979620b3',
+                category: 'risk',
+              },
+              {
+                key: 'health_issue',
+                title: 'Health issue',
+                comments: 'Keeps complaining of headaches',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: '5bb4f1a2-b5e3-49af-a2ef-01b0ae33429d',
+                category: 'health',
+              },
+              {
+                key: 'other_risks',
+                title: 'Any other risks',
+                comments: '',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: '94cf7afd-d2c7-489f-b6c2-a6fb68cc9f15',
+                category: 'risk',
+              },
+              {
+                key: 'legal_representation',
+                title: 'Solicitor or other legal representation',
+                comments: '',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: 'f191fe25-11c8-4195-bbcc-99bf508f293d',
+                category: 'court',
+              },
+              {
+                key: 'invalid_answer',
+                title: 'Invalid answer',
+                comments: '',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: 'f1f1fe15-11c8-4195-bbcc-99bf508f293d',
+                category: 'invalid',
+              },
+              {
+                key: 'escape',
+                title: 'Escape',
+                comments: 'Large poster in cell',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: 'f199c4fe-0134-490c-afa9-a11f3c52c32b',
+                category: 'risk',
+              },
+              {
+                key: 'diet_allergy',
+                title: 'Special diet or allergy',
+                comments: 'Vegan',
+                date: null,
+                expiry_date: null,
+                assessment_question_id: '394d3f05-4d25-43ef-8e7e-6d3a0e742888',
+                category: 'health',
+              },
+            ],
+          }
+
+          transformedResponse = personToCardComponent()(mockPersonWithAnswers)
+        })
+
+        it('should correctly filter', function() {
+          expect(transformedResponse).to.have.property('tags')
+          expect(transformedResponse.tags.items.length).to.equal(5)
+        })
+
+        it('should correctly map and sort', function() {
+          expect(transformedResponse.tags.items).to.deep.equal([
+            {
+              href: `${mockPersonWithAnswers.href}#concealed-items`,
+              text: 'Concealed items',
+              classes: 'app-tag--destructive',
+              sortOrder: 1,
+            },
+            {
+              href: `${mockPersonWithAnswers.href}#any-other-risks`,
+              text: 'Any other risks',
+              classes: 'app-tag--destructive',
+              sortOrder: 1,
+            },
+            {
+              href: `${mockPersonWithAnswers.href}#escape`,
+              text: 'Escape',
+              classes: 'app-tag--destructive',
+              sortOrder: 1,
+            },
+            {
+              href: `${mockPersonWithAnswers.href}#health-issue`,
+              text: 'Health issue',
+              classes: '',
+              sortOrder: 2,
+            },
+            {
+              href: `${mockPersonWithAnswers.href}#special-diet-or-allergy`,
+              text: 'Special diet or allergy',
+              classes: '',
+              sortOrder: 2,
+            },
+          ])
+        })
+      })
+    })
+
+    context('with meta disabled', function() {
+      beforeEach(function() {
+        transformedResponse = personToCardComponent({
+          showMeta: false,
+        })(mockPerson)
+      })
+
+      it('should not contain meta items', function() {
+        expect(transformedResponse).to.have.property('meta')
+        expect(transformedResponse.meta.items).to.be.undefined
+      })
+
+      it('should contain href', function() {
+        expect(transformedResponse).to.have.property('href')
+      })
+
+      it('should contain title', function() {
+        expect(transformedResponse).to.have.property('title')
+      })
+
+      it('should contain tags', function() {
+        expect(transformedResponse).to.have.property('tags')
+        expect(transformedResponse.tags.items.length).to.equal(3)
+      })
+    })
+
+    context('with tag disabled', function() {
+      beforeEach(function() {
+        transformedResponse = personToCardComponent({
+          showTags: false,
+        })(mockPerson)
+      })
+
+      it('should not contain tags items', function() {
+        expect(transformedResponse).to.have.property('tags')
+        expect(transformedResponse.tags.items).to.be.undefined
+      })
+
+      it('should contain href', function() {
+        expect(transformedResponse).to.have.property('href')
+      })
+
+      it('should contain title', function() {
+        expect(transformedResponse).to.have.property('title')
+      })
+
+      it('should contain meta', function() {
+        expect(transformedResponse).to.have.property('meta')
+        expect(transformedResponse.meta.items.length).to.equal(2)
       })
     })
   })
